streamlit_app/requirements.txt
streamlit>=1.27.0
llama-index>=0.9.0
llama-index-vector-stores-pinecone>=0.1.4
llama-index-embeddings-voyageai>=0.1.3
llama-index-llms-anthropic>=0.1.3
llama-index-readers-google>=0.1.3
anthropic>=0.8.0
pinecone-client>=2.2.4
python-dotenv>=1.0.0
tqdm>=4.66.1